рефакторинг и реализация получения токена для дальнейшего использования в запросах.
