+++ app/src/main/java/com/muslim/simplevknewsclient/presentation/comments/CommentsScreenState.kt
@@ -1,4 +1,4 @@
-package com.muslim.simplevknewsclient
+package com.muslim.simplevknewsclient.presentation.comments
 
 import com.muslim.simplevknewsclient.domain.FeedPost
 import com.muslim.simplevknewsclient.domain.PostComment
