+++ app/src/main/java/com/muslim/simplevknewsclient/presentation/comments/CommentsViewModelFactory.kt
@@ -1,4 +1,4 @@
-package com.muslim.simplevknewsclient
+package com.muslim.simplevknewsclient.presentation.comments
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
