+++ app/src/main/java/com/muslim/simplevknewsclient/presentation/main/AuthState.kt
@@ -1,4 +1,4 @@
-package com.muslim.simplevknewsclient
+package com.muslim.simplevknewsclient.presentation.main
 
 sealed class AuthState {
