+++ app/src/main/java/com/muslim/simplevknewsclient/presentation/main/NavigationItem.kt
@@ -1,10 +1,11 @@
-package com.muslim.simplevknewsclient
+package com.muslim.simplevknewsclient.presentation.main
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Favorite
 import androidx.compose.material.icons.outlined.Home
 import androidx.compose.material.icons.outlined.Person
 import androidx.compose.ui.graphics.vector.ImageVector
+import com.muslim.simplevknewsclient.R
 import com.muslim.simplevknewsclient.navigation.Screen
 
 sealed class NavigationItem(
