+++ app/src/main/java/com/muslim/simplevknewsclient/presentation/main/VkNewsMainScreen.kt
@@ -1,4 +1,4 @@
-package com.muslim.simplevknewsclient
+package com.muslim.simplevknewsclient.presentation.main
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.clickable
@@ -10,10 +10,8 @@ import androidx.compose.material3.NavigationBarItemDefaults
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
@@ -21,10 +19,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.navigation.NavDestination.Companion.hierarchy
 import androidx.navigation.compose.currentBackStackEntryAsState
-import com.muslim.simplevknewsclient.domain.FeedPost
 import com.muslim.simplevknewsclient.navigation.AppNavGraph
-import com.muslim.simplevknewsclient.navigation.Screen
 import com.muslim.simplevknewsclient.navigation.rememberNavigationState
+import com.muslim.simplevknewsclient.presentation.comments.CommentsScreen
+import com.muslim.simplevknewsclient.presentation.news.NewsFeedScreen
 
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @Composable
@@ -74,7 +72,7 @@ fun MainScreen() {
         AppNavGraph(
             navHostController = navigateState.navHostController,
             newsFeedScreenContent = {
-                HomeScreen(onCommentClickListener = {
+                NewsFeedScreen(onCommentClickListener = {
                     navigateState.navigateToComments(it)
                 })
             },
