+++ app/src/main/java/com/muslim/simplevknewsclient/presentation/news/NewsFeedViewModel.kt
@@ -1,4 +1,4 @@
-package com.muslim.simplevknewsclient
+package com.muslim.simplevknewsclient.presentation.news
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
